Publish toPyPI on release.
This uses a modified version of the workflow from
https://github.com/actions/starter-workflows/blob/main/ci/python-publish.yml

In particular, we use the trusted publisher workflow. See
https://docs.github.com/en/actions/deployment/security-hardening-your-deployments/about-security-hardening-with-openid-connect#adding-permissions-settings

Currently there is no (automatic) link between the release version, as
set by a tag, and the version in the pyproject.toml file. Thus it is
important to ensure these match when making a release.