Ensure rows added for only selected foreign-keys.
Previously, when building the graph, the code would only add rows
realted through a reverse foreign-key relation only if they were
from a table included in the `only_tables` argument. However, rows
related by forward foreign-keys would be added regardless. This was a
mistake, and now the forward case is similar to the reverse case.